pyproject.toml
[project]
name = "simple-rag-agent"
version = "0.1.0"
description = "Simple RAG agent with Image Captioning using ChromaDB and BLIP model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "uvicorn>=0.34.2",
    "httpx>=0.28.1",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "pillow>=10.0.0",
    "google-genai>=1.9.0",
    "google-adk>=1.7.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "oci>=2.0.0",
    "langchain-community>=0.0.20",
    "langchain-core>=0.1.0",
    "langchain-oci>=0.1.0",
    "pyyaml>=6.0.0",
    "scikit-learn>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

